+++ aa_engine/aas_processBIDS.m
@@ -63,7 +63,7 @@ SUBJ = spm_select('List',aap.directory_conventions.rawdatadir,'dir','sub-.*');
 % 1st pass - Add sessions only
 % 2ns pass - Add data
 for p = [false true]
-    for subj = 1%:size(SUBJ,1)
+    for subj = 1:size(SUBJ,1)
         subjID = deblank(SUBJ(subj,:));
         SESS = spm_select('List',fullfile(BIDS,subjID),'dir','ses');
         if isempty(SESS)
@@ -140,6 +140,8 @@ for cf = cellstr(spm_select('List',sesspath,'dir'))'
 
                     if ~isempty(strfind(basename(sesspath),'ses-'))
                         sesstr = ['_' strrep(basename(sesspath),'ses-','')];
+                    else
+                        sesstr = '';
                     end
                     
                     if BIDSsettings.combinemultiple
@@ -152,7 +154,7 @@ for cf = cellstr(spm_select('List',sesspath,'dir'))'
                     if ~toAddData, continue; end
                     
                     aasessnames = {aap.acq_details.diffusion_sessions.name};
-                    if BIDSsettings.combinemultiple
+                    if BIDSsettings.combinemultiple && ~ isempty(sesstr)
                         aasessnames = aasessnames(cell_index(aasessnames,sesstr));
                     end
                     if isempty(diffusionimages), diffusionimages = cell(1,numel(aasessnames)); end
@@ -211,7 +213,7 @@ for cf = cellstr(spm_select('List',sesspath,'dir'))'
                 
                 % Data
                 aasessnames = {aap.acq_details.sessions.name};
-                if BIDSsettings.combinemultiple
+                if BIDSsettings.combinemultiple && ~isempty(sesssfx)
                     aasessnames = aasessnames(cell_index(aasessnames,sesssfx));
                 end
                 if isempty(functionalimages), functionalimages = cell(1,numel(aasessnames)); end
@@ -338,8 +340,8 @@ function sessord = sort_sessions(sessions,sessstr)
 aasessnames = {sessions.name};
 sessstr = spm_file(sessstr,'prefix','_');
 sessord = [];
-for sess = sessstr
-    sessord = horzcat(sessord, cell_index(aasessnames,sess{1}));
+for sess = sessstr'
+    sessord = horzcat(sessord, cell_index(aasessnames,sess{1})');
 end
 end