+++ aa_modules/aamod_dartel_denorm.m
@@ -16,10 +16,11 @@ switch task
         inpstreams = aas_getstreams(aap,'input');
         workstream = inpstreams{end};
         images = cellstr(aas_getfiles_bystream_multilevel(aap, domain, [subj sess], workstream));
-        regstreams = inpstreams(1:end-3);
+        regstreams = inpstreams(1:end-3); toRemove = [];
         for i = 1:numel(regstreams)
-            if ~aas_stream_has_contents(aap,domain,[subj sess],regstreams{i}), regstreams(i) = []; end
+            if ~aas_stream_has_contents(aap,domain,[subj sess],regstreams{i}), toRemove(end+1) = i; end
         end
+        regstreams(toRemove) = [];
 
         streamfn = aas_getfiles_bystream(aap,domain,[subj sess],['native_' workstream],'output');
         streamfn = strtok_ptrn(basename(streamfn(1,:)),'-0');
@@ -45,13 +46,14 @@ switch task
         inpstreams = aas_getstreams(aap,'input');
         workstream = inpstreams{end};
         images = cellstr(aas_getfiles_bystream_multilevel(aap, domain, [subj sess], workstream));
-        regstreams = inpstreams(1:end-3);
+        regstreams = inpstreams(1:end-3); toRemove = [];
         for i = 1:numel(regstreams)
-            if ~aas_stream_has_contents(aap,domain,[subj sess],regstreams{i}), regstreams(i) = []; end
+            if ~aas_stream_has_contents(aap,domain,[subj sess],regstreams{i}), toRemove(end+1) = i; end
         end
+        regstreams(toRemove) = [];
 
         % Affine Template2mni
-        xfm = load(aas_getfiles_bystream_multilevel(aap, subj, sess, 'dartel_templatetomni_xfm'));
+        xfm = load(aas_getfiles_bystream_multilevel(aap, domain, [subj sess], 'dartel_templatetomni_xfm'));
         for i = 1:numel(images)
             if ~exist(fullfile(aas_getpath_bydomain(aap,domain,[subj,sess]),[basename(images{i}) '.nii']),'file')
                 copyfile(images{i},fullfile(aas_getpath_bydomain(aap,domain,[subj,sess]),[basename(images{i}) '.nii']));